deviceNodeServer/DeviceMainNodeServer/VideoHandler_v2/requirements.txt
aiomysql==0.3.2
gen==0.1.30
numpy==1.26.4
opencv-python==4.6.0.66
pyzmq>=25.1.0
tornado==6.0.4
mysql-connector-python==8.0.33